// src/components/ActionButton.d.ts
import { ThemeMode } from '@/styles/theme_colors';
declare const ActionButton: import("styled-components/dist/types").IStyledComponentBase<"web", import("styled-components/dist/types").Substitute<import("react").DetailedHTMLProps<import("react").ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement>, {
    $themeMode: ThemeMode;
}>> & string;
export default ActionButton;